pyproject.toml
[project]
name = "PythonBasics"
version = "0.1.0"
description = "-- --"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "box>=0.1.5",
    "httpx>=0.28.1",
    "ollama>=0.4.8",
    "pypdf2>=3.0.1",
    "python-box>=7.3.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
